pyproject.toml
[project]
name = "bayesian-quiz"
version = "0.1.0"
description = "Meetup quiz where you quantify your uncertainty"
readme = "README.md"
authors = [
    { name = "Jouni K. Seppänen", email = "jks@iki.fi" }
]
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.127.1",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.21",
    "qrcode[pil]>=8.2",
    "uvicorn>=0.40.0",
]

[project.scripts]
bayesian-quiz = "bayesian_quiz:main"

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "anyio[trio]>=4.12.0",
    "pytest>=9.0.2",
    "pytest-playwright>=0.7.2",
    "ruff>=0.15.2",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes (unused imports, undefined names)
    "I",     # isort
    "B",     # flake8-bugbear (suspicious constructs, raise-without-cause)
    "UP",    # pyupgrade (modern Python idioms)
    "ASYNC", # flake8-async (async best practices)
    "PT",    # flake8-pytest-style
    "FAST",  # FastAPI-specific rules
    "RUF",   # ruff-specific rules
]
ignore = [
    "RUF002",  # Greek/math symbols in docstrings are intentional in this codebase
]

[tool.ruff.lint.per-file-ignores]
"tests/test_xss.py" = ["E501"]  # XSS payload strings cannot be shortened
